moving socket to a different file.
